fix(IO): add explicit casts in `HistogramMetrics::observe` calls
`observe` takes `double` but `elapsedMicroseconds` returns `UInt64`
and byte counts are `size_t`. The implicit conversion triggers
`-Werror,-Wimplicit-int-float-conversion` on macOS builds after
merging master.

+++ src/IO/S3/ReadBufferFromGetObjectResult.cpp
@@ -49,8 +49,8 @@ void ReadBufferFromGetObjectResult::observeMetrics()
         return;
     metrics_observed = true;
 
-    HistogramMetrics::S3ReadRequestDuration.observe(watch.elapsedMicroseconds());
-    HistogramMetrics::S3ReadRequestBytes.observe(bytes_read);
+    HistogramMetrics::S3ReadRequestDuration.observe(static_cast<double>(watch.elapsedMicroseconds()));
+    HistogramMetrics::S3ReadRequestBytes.observe(static_cast<double>(bytes_read));
 }
 
 }
